Update Anchor usage
Extract Anchor into a more specific component.
Enables the creation of custom Anchors.
