+++ src/atoms/anchor/index.jsx
@@ -1,45 +1,7 @@
-import React from 'react';
+import Anchor from './regularAnchor';
 import { Link } from 'gatsby';
-import PropTypes from 'prop-types';
-import { Link as LinkPropType } from 'proptypes';
 
-const Anchor = ({
-  children,
+export {
+  Anchor,
   link,
-  ...rest
-}) => {
-  return link.internal ?
-    (
-      <Link
-        to={ link.url }
-        rel={ link.rel }
-        target={ link.target }
-        { ...rest }
-      >
-        { children }
-      </Link>
-    ) : (
-      <a
-        href={ link.url }
-        rel={ link.rel }
-        target={ link.target }
-        { ...rest }
-      >
-        { children }
-      </a>
-    );
 };
-
-Anchor.propTypes = {
-  /** Children elements */
-  children: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node),
-  ]),
-  /** Anchor link data */
-  link: LinkPropType.isRequired,
-  /** Any other props to be passed to the component */
-  rest: PropTypes.any,
-};
-
-export default Anchor;
