+++ src/atoms/anchor/regularAnchor/docs.mdx
@@ -8,7 +8,7 @@ import {
   ColorItem,
   Meta,
 } from '@storybook/addon-docs/blocks';
-import Anchor from 'atoms/anchor';
+import { Anchor } from 'atoms/anchor';
 
 # Anchor
